requirements.txt
numpy==2.0.0
pandas==2.2.2
PyYAML==6.0.2
requests==2.32.3
yfinance==0.2.40
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.1
imageio==2.34.1
imageio-ffmpeg==0.4.9
tqdm==4.66.4
pandas-datareader==0.10.0
